+++ packages/zosfiles/src/methods/list/List.ts
@@ -48,14 +48,18 @@ export class List {
 
         try {
             // Format the endpoint to send the request to
-            let endpoint = posix.join(
+            const endpoint = posix.join(
                 ZosFilesConstants.RESOURCE,
                 ZosFilesConstants.RES_DS_FILES,
                 encodeURIComponent(dataSetName),
                 ZosFilesConstants.RES_DS_MEMBERS);
 
+            const params = new URLSearchParams();
             if (options.pattern) {
-                endpoint += `?pattern=${encodeURIComponent(options.pattern)}`;
+                params.set("pattern", options.pattern);
+            }
+            if (options.start) {
+                params.set("start", options.start);
             }
 
             const reqHeaders: IHeaderContent[] = [ZosmfHeaders.ACCEPT_ENCODING];
@@ -73,7 +77,7 @@ export class List {
 
             this.log.debug(`Endpoint: ${endpoint}`);
 
-            const data = await ZosmfRestClient.getExpectString(session, endpoint, reqHeaders);
+            const data = await ZosmfRestClient.getExpectString(session, endpoint.concat(params.size > 0 ? `?${params.toString()}` : ""), reqHeaders);
             let response: any;
             try {
                 response = JSONUtils.parse(data);
@@ -124,7 +128,7 @@ export class List {
         const zosmfResponses: IZosmfListResponse[] = [];
 
         for(const pattern of patterns) {
-            const response = await List.allMembers(session, dataSetName, { pattern});
+            const response = await List.allMembers(session, dataSetName, { pattern, maxLength: options.maxLength, start: options.start });
             zosmfResponses.push(...response.apiResponse.items);
         }
 
@@ -139,7 +143,7 @@ export class List {
 
         // Exclude names of members
         for (const pattern of options.excludePatterns || []) {
-            const response = await List.allMembers(session, dataSetName, {pattern});
+            const response = await List.allMembers(session, dataSetName, { pattern });
             response.apiResponse.items.forEach((membersObj: IZosmfListResponse) => {
                 const responseIndex = zosmfResponses.findIndex(response=> response.member === membersObj.member);
                 if (responseIndex !== -1) {
@@ -426,7 +430,7 @@ export class List {
         for (const pattern of patterns) {
             let response: any;
             try {
-                response = await List.dataSet(session, pattern, { attributes: true });
+                response = await List.dataSet(session, pattern, { attributes: true, maxLength: options.maxLength, start: options.start });
             } catch (err) {
                 if (!(err instanceof ImperativeError && err.errorCode?.toString().startsWith("5"))) {
                     throw err;